[WIP] workflows: run tag and build in one workflow
GitHub workflows are limited in that a workflow (here: tag) can not
trigger another workflow (here: build).

Both worked independently, yet the build workflow running "on" tag
"create" would only run upon push, not when the other worfklow created a
tag via the GitHub script.

To avoid managing personal access tokens, which would allow this, the
actions are merged into one.

The tag name is change to be more generic.

https://docs.github.com/en/actions/writing-workflows/choosing-when-your-workflow-runs/triggering-a-workflow#triggering-a-workflow-from-a-workflow